Moved DeviceLimits into it's own cpp file.
