+++ CMakeLists.txt
@@ -46,8 +46,8 @@ endif()
 set(SRC_DIR ${CMAKE_CURRENT_SOURCE_DIR}/src)
 set(SRC_FILES
     ${SRC_DIR}/particle_io.cpp ${SRC_DIR}/cell_dat_compression.cpp
-    ${SRC_DIR}/parallel_initialisation.cpp ${SRC_DIR}/particle_group.cpp
-    ${SRC_DIR}/cell_dat_move.cpp)
+    ${SRC_DIR}/cell_dat_move.cpp ${SRC_DIR}/device_limits.cpp
+    ${SRC_DIR}/parallel_initialisation.cpp ${SRC_DIR}/particle_group.cpp)
 set(SRC_FILES_IGNORE "")
 check_added_file_list(${SRC_DIR} cpp "${SRC_FILES}" "${SRC_FILES_IGNORE}")
